Move PID to separate module
